Moves test to different fixture
